Claude's phase 1 refactor of XIP code. I have tested it builds, not the functionality yet.
