+++ EPII_CM55M_APP_S/app/ww_projects/ww500_md/cvapp.cpp
@@ -29,6 +29,7 @@
 #include "cvapp.h"
 #include "ff.h"
 
+// SPI includes retained for the #if SPIDEBUG block below
 #include "hx_drv_spi.h"
 #include "spi_eeprom_comm.h"
 
@@ -53,16 +54,12 @@
 #include "printf_x.h" // Print colours
 #include "ww500_md.h"
 #include "app_msg.h"
+#include "xip_manager.h"
 
 /*************************************** Definitions *******************************************/
 
 // Experiment to add multiple resolvers to NN
 #define EXTRARESOLVERS
-
-// Read file contents in chunks this size
-#define FILE_CHUNK_SIZE 512
-
-// #include "dev_common.h"
 
 #define LOCAL_FRAQ_BITS (8)
 #define SC(A, B) ((A << 8) / B)
@@ -80,21 +77,14 @@
 // Uncomment this to get information about the model:
 #define PRINTMODELFINGERPRINT
 
-// oRIGINALLY, SOME META DATA WAS STORED THIS NUMBER OF BYTES BEFORE THE MODEL
-#define MODEL_OLD_META_SIZE	24
+// #define OLD
 
-// Metadata stored in XIP flash just before the model itself.
-typedef struct {
-    uint32_t magic;
-    uint16_t class_count;
-    uint16_t label_len;
-    char modelName[MAX_MODEL_NAME_LEN];
-    char labels[MAX_CLASSES][MAX_LABEL_LEN];
-    uint32_t crc;   // optional but highly recommended
-} ModelMetaData;
-
-// This version becomes available if a valid structure is written to flash - only at the virtual address 0x3A200000
-#define metaDataFlash ((const ModelMetaData *)MODEL_XIP_ADDR)
+#ifdef TFLM_2412
+//	Later library - defined in the makefile
+#pragma message "Using tflmtag2412_u55tag2411"
+#else
+#pragma message "Using tflmtag2209_u55tag2205"
+#endif // TFLM_2412
 
 /*************************************** External variables *******************************************/
 
@@ -112,17 +102,8 @@ extern QueueHandle_t xImageTaskQueue;
 
 /*************************************** C++ Namespace *******************************************/
 
-// #define OLD
 // see https://chatgpt.com/share/696ae627-6f60-8005-92cb-6030e56990cc
 // for fix of dynamic model
-
-
-#ifdef TFLM_2412
-//	Later library - defined in the makefile
-#pragma message "Using tflmtag2412_u55tag2411"
-#else
-#pragma message "Using tflmtag2209_u55tag2205"
-#endif // TFLM_2412
 
 using namespace std;
 namespace {
@@ -140,19 +121,10 @@ namespace {
     TfLiteTensor *input;
     TfLiteTensor *output;
 
-    static uint8_t g_label_count = 0;
-    static uint8_t g_class_count = 0;	// These should be teh same
+    static uint8_t g_class_count = 0;
 };
 
-/*************************************** Local variables *******************************************/
-
-// Use this constants since a bizarre compiler issue is redefining USE_DW_SPI_MST_Q as 1860 in some places
-// See the "Bizarre" comment in is_model_in_flash()
-USE_DW_SPI_MST_E spi_inst = (USE_DW_SPI_MST_E)0;
-
-// SPI EEPROM configuration
-static bool flash_initialized = false;
-static SemaphoreHandle_t xSPIMutex = NULL;
+/*************************************** Local variables *************************************/
 
 // Globals to hold the current model identifiers
 // TODO should these really be 32-bits?
@@ -168,54 +140,15 @@ static bool coldBoot;
 
 /*************************************** Local Function Declarations *****************************/
 
-int xip_init_flash(void);
-
-// not used static inline uint32_t align_down(uint32_t addr, uint32_t align);
-static inline uint32_t xip_align_up(uint32_t size, uint32_t align);
-
-static bool xip_is_model_in_flash(char *filename);
-static bool xip_is_file_in_sd(char *filename);
-static const tflite::Model *xip_load_model_from_sd(char * filename);
-
+static const tflite::Model *xip_load_model_from_sd(char *filename);
 static const tflite::Model *xip_load_model_from_flash(void);
 
-int xip_load_model_from_sd_to_flash(char *filename);
-
-int xip_erase_model_flash_area(uint32_t flashSizeRequired);
-
-
-static bool xip_copy_model_from_sd_to_flash(char * filename);
-
-static bool xip_copy_metadata_to_flash(char * modelName);
-
-static bool xip_valid_model_in_flash(void);
-
-int32_t xip_write_meta_data_to_flash(ModelMetaData * metaDataRam);
-
-
 #ifdef SPIDEBUG
-//Not used functions
-
 void compare_sd_spi_xip_chunked(const char *filename);
 void debug_sd_model_integrity(const char *filename);
 void debug_flash_status(void);
 void test_basic_spi_operations();
 #endif // SPIDEBUG
-
-
-static bool file_exists(const char *path);
-static uint8_t load_labels_from_manifest(char * filename, char (*labels)[MAX_LABEL_LEN]);
-
-
-static bool xip_enable_XIP(bool enable);
-
-static inline uint32_t xip_virt_to_phys(uint32_t virt)  {
-	return virt - FLASH_VIRTUAL_BASE + FLASH_PHYSICAL_BASE;
-}
-
-static inline uint32_t xip_phys_to_virt(uint32_t phys) {
-	return phys - FLASH_PHYSICAL_BASE + FLASH_VIRTUAL_BASE;
-}
 
 #ifdef USE_PERCENTAGE
 static void outputAsPercentage(TfLiteTensor *output);
@@ -224,97 +157,6 @@ static void outputAsPercentage(TfLiteTensor *output);
 static void tflm_print_ethosu_fingerprint(const tflite::Model* model);
 
 /*************************************** Local Function Definitions  *************************************/
-
-// Not used
-//// Round down address to sector boundary
-//static inline uint32_t align_down(uint32_t addr, uint32_t align) {
-//    return addr & ~(align - 1);
-//}
-
-/**
- * Rounds up 'size' to be an integral number of 'align' bytes
- *
- * Examples if 'align' is 4:
- *
- * size = 0: 0+4-1 = 3  then 0011b & 0011b = 0
- * size = 1: 1+4-1 = 4  then 0100b & 0011b = 0100b = 4
- * size = 2: 2+4-1 = 5  then 0101b & 0011b = 0100b = 4
- * size = 3: 3+4-1 = 6  then 0110b & 0011b = 0100b = 4
- * size = 4: 4+4-1 = 7  then 0111b & 0011b = 0100b = 4
- * size = 5: 5+4-1 = 8  then 1000b & 0011b = 1000b = 8
- *
- * @param size - number being examined
- * @param align - number of bytes in a chunk
- * @return - a number which is greater or equal to 'size' and
- */
-static inline uint32_t xip_align_up(uint32_t size, uint32_t align) {
-    return (size + align - 1) & ~(align - 1);
-}
-
-// Check if a file exists on SD
-static bool file_exists(const char *path) {
-    FILINFO finfo;
-    FRESULT res = f_stat(path, &finfo);
-    return (res == FR_OK);
-}
-
-/**
- * Load labels from "/MANIFEST/nnVvv.TXT"
- *
- * The parameter is the model name (ends with '.TFL') and the labels are expected in a file
- * with the same base name but with .TXT ending.
- *
- * If the file exists then the labels are loaded into the metaDataRam structure
- * the number of labels is in g_label_count
- *
- * @param filename- pointer to the string containing the model name
- * @param labels - pointer to the metaDataRam structure, labels section
- * return - number of labels in file (0 if no file etc)
- */
-static uint8_t load_labels_from_manifest(char * filename, char (*labels)[MAX_LABEL_LEN]) {
-
-	char labels_path[MAX_MODEL_NAME_LEN * 2];   // plenty for "/MANIFEST/xxxxxxxx.TXT"
-	char base[MAX_MODEL_NAME_LEN];          // 8.3 name + NUL
-
-	// Copy filename so we can modify it
-	strncpy(base, filename, sizeof(base) - 1);
-
-	base[sizeof(base) - 1] = '\0';
-
-	/* Replace extension */
-	char *dot = strrchr(base, '.');
-	if (dot != NULL) {
-		strcpy(dot, ".TXT");
-	}
-	else {
-		// Should not happen
-		return 0;
-	}
-
-	/* Build full path */
-	snprintf(labels_path, sizeof(labels_path), "%s/%s", CONFIG_DIR, base);
-
-	xprintf("Looking for labels in %s\n", labels_path);
-
-	if (file_exists(labels_path))  {
-		uint8_t count = 0;
-		// TODO - labels to flash
-        // https://chatgpt.com/share/696d4dca-4900-8005-a0ad-9a211e8a1d9a
-        if (fatfs_load_labels(labels_path, labels, &count, MAX_CLASSES, MAX_LABEL_LEN) == 0) {
-            g_label_count = count;
-            xprintf("Loaded %d labels from %s\n", g_label_count, labels_path);
-            return g_label_count;
-        }
-        else  {
-            xprintf("Labels load failed from %s\n", labels_path);
-            return 0;
-        }
-    }
-	else {
-		xprintf("File %s not found\n", labels_path);
-		return 0;
-	}
-}
 
 #ifdef PRINTMODELFINGERPRINT
 #include "tensorflow/lite/schema/schema_generated.h"
@@ -534,7 +376,6 @@ static int _arm_npu_init(bool security_enable, bool privilege_enable) {
     _arm_npu_irq_init();
 
     /* Initialise Ethos-U55 device */
-    //const void *ethosu_base_address = (void *)(U55_BASE);
     void *ethosu_base_address = (void *)(U55_BASE);
 
     if (0 != (err = ethosu_init(
@@ -555,10 +396,11 @@ static int _arm_npu_init(bool security_enable, bool privilege_enable) {
 }
 
 
-// Not used
+// Not used — debug helpers kept here under SPIDEBUG guard.
+// NOTE: these reference spi_inst which has moved to xip_manager.c.
+// This block is intentionally dead code (#if SPIDEBUG evaluates to #if 0).
+// TODO Phase 2: remove or relocate this block.
 #if SPIDEBUG
-
-/* ------------debuggers------------- */
 
 void debug_sd_model_integrity(const char *filename) {
     xprintf("=== SD MODEL INTEGRITY CHECK ===\n");
@@ -599,19 +441,12 @@ void debug_sd_model_integrity(const char *filename) {
 }
 
 
-/**
- * Debug code for SPI EEPROM:
- *
- * Prints Flash init status, SPI ID, EEPROM ID byte
- */
 void debug_flash_status(void) {
     uint8_t id_info = 0;
     int result;
 
-    //xprintf("=== FLASH DEBUG STATUS ===\n");
     xprintf("Flash initialized: %s\n", flash_initialized ? "YES" : "NO");
 
-    // Disable XIP before reading ID
     xip_enable_XIP(false);
 
     if (xSemaphoreTake(xSPIMutex, portMAX_DELAY) != pdTRUE)  {
@@ -641,7 +476,6 @@ void test_basic_spi_operations() {
 
     uint32_t test_data[4] = {0xDEADBEEF, 0x12345678, 0xABCDEF00, 0x11223344};
 
-    // Test writing one word at a time
     xprintf("Writing one word at a time:\n");
     for (int i = 0; i < 4; i++) {
         int wr = hx_lib_spi_eeprom_word_write(spi_inst, test_addr + (i * 4), &test_data[i], 1);
@@ -649,7 +483,6 @@ void test_basic_spi_operations() {
                 i, test_data[i], test_addr + (i * 4), wr);
     }
 
-    // Test reading one word at a time
     xprintf("Reading one word at a time:\n");
     for (int i = 0; i < 4; i++)  {
         uint32_t read_data = 0;
@@ -659,11 +492,7 @@ void test_basic_spi_operations() {
                 (read_data == test_data[i]) ? "OK" : "FAIL");
     }
 }
-*/
 
-
-
-// Not used
 void compare_sd_spi_xip_chunked(const char *filename) {
     FIL file;
     FRESULT res = f_open(&file, filename, FA_READ);
@@ -686,7 +515,6 @@ void compare_sd_spi_xip_chunked(const char *filename) {
     int any_mismatch = 0;
     while (offset < fileSize)  {
         UINT to_read = (fileSize - offset > FILE_CHUNK_SIZE) ? FILE_CHUNK_SIZE : (fileSize - offset);
-        // Read SD
         f_lseek(&file, offset);
         res = f_read(&file, sd_buf, to_read, &bytesRead);
         if (res != FR_OK || bytesRead != to_read)
@@ -694,18 +522,14 @@ void compare_sd_spi_xip_chunked(const char *filename) {
             xprintf("Failed to read model file at offset %lu\n", offset);
             break;
         }
-        // // Read SPI
         if (hx_lib_spi_eeprom_word_read(spi_inst, MODEL_FLASH_ADDR + offset, (uint32_t *)spi_buf, to_read) != 0)
         {
-            // if (hx_lib_spi_eeprom_word_read(spi_inst, MODEL_FLASH_ADDR + offset, (uint32_t*)spi_buf, (to_read + 3) / 4) != 0) {
             xprintf("Failed to read SPI at offset %lu\n", offset);
             break;
         }
-        // Read XIP
         uint32_t virt_address = xip_phys_to_virt(MODEL_FLASH_ADDR + offset);
         memcpy(xip_buf, (void *)virt_address, to_read);
 
-        // Print mismatches for this chunk
         for (UINT i = 0; i < to_read; i++)
         {
             if (sd_buf[i] != spi_buf[i] || sd_buf[i] != xip_buf[i] || spi_buf[i] != xip_buf[i])
@@ -731,9 +555,10 @@ void compare_sd_spi_xip_chunked(const char *filename) {
  *
  * Model is expected to be in /MANIFEST folder.
  *
- * Model is copied to flash and a pointer to the model is returned.
+ * Model and metadata are copied to flash, then a pointer to the model
+ * is returned via xip_load_model_from_flash().
  */
-static const tflite::Model *xip_load_model_from_sd(char * filename) {
+static const tflite::Model *xip_load_model_from_sd(char *filename) {
 
 	if (xip_copy_model_from_sd_to_flash(filename)) {
 		xprintf("Copied %s to flash OK\n", filename);
@@ -754,886 +579,15 @@ static const tflite::Model *xip_load_model_from_sd(char * filename) {
 	return xip_load_model_from_flash();
 }
 
-
 /**
- * Initialize the flash memory (SPI EEPROM)
+ * Validate the model in flash and return a TFLite model pointer.
  *
- * TODO - this is called many times - should only be called once.
- */
-int xip_init_flash(void) {
-	uint8_t id_info;
-
-	if (flash_initialized) {
-		return 0;
-	}
-
-	//xprintf("Init EEPROM...\r\n");
-
-	// Create mutex on first call
-	// TODO for consistenacy this should be in image_createTask()
-	// semaphore vs mutex: https://chatgpt.com/share/69706528-d250-8005-973b-6ab43c1b4629
-	// In FreeRTOS, a mutex is created in the given (unlocked) state.
-	if (xSPIMutex == NULL)  {
-		xSPIMutex = xSemaphoreCreateMutex();
-		if (xSPIMutex == NULL) {
-			xprintf("Failed to create SPI mutex\n");
-			return -1;
-		}
-	}
-
-	//	hx_lib_spi_eeprom_open(spi_inst);
-	//    //hx_lib_spi_eeprom_open_speed(spi_inst, 6000000);
-	//
-	//	hx_lib_spi_eeprom_enable_XIP(spi_inst, true, FLASH_QUAD, true);
-
-	// Take mutex
-	if (xSemaphoreTake(xSPIMutex, portMAX_DELAY) != pdTRUE) {
-		xprintf("Failed to take SPI mutex\n");
-		return -1;
-	}
-
-	if (hx_lib_spi_eeprom_open(spi_inst) != 0) {
-		xprintf("Failed to open SPI EEPROM\n");
-		flash_initialized = false;
-		xSemaphoreGive(xSPIMutex);
-		return -1;
-	}
-
-	// Small delay for mode switch
-	// TODO - why?
-	vTaskDelay(pdMS_TO_TICKS(10));
-
-	if (hx_lib_spi_eeprom_read_ID(spi_inst, &id_info) != 0) {
-		xprintf("Failed to read flash ID\n");
-		xSemaphoreGive(xSPIMutex);
-		return -1;
-	}
-
-	// Disable XIP mode, so that SPI access are possible
-	if (hx_lib_spi_eeprom_enable_XIP(spi_inst, false, FLASH_QUAD, true) != 0) {
-		xprintf("Failed to disable XIP mode\n");
-		xSemaphoreGive(xSPIMutex);
-		return -1;
-	}
-
-	xSemaphoreGive(xSPIMutex);
-
-	xprintf("Flash ID 0x%02X Initialised\n", id_info);
-	flash_initialized = true;
-
-	return 0;
-}
-
-// Erase the flash region covering the model info, header, AND the model data.
-/**
- * Erase an area of flash for the model file plus meta information.
- * The meta information will be placed first.
- *
- * Flash is erased in chunks of 4k bytes
- *
- * @param model_size = size of model file read from SD card
- */
-int xip_erase_model_flash_area(uint32_t flashSizeRequired) {
-    int32_t ret = 0;
-
-    uint32_t sectors_needed;
-	uint32_t sector_addr;
-
-	if (xip_init_flash() != 0) {
-		return -1;
-	}
-
-    // Calculate the number of 4k sectors needed
-    sectors_needed = xip_align_up(flashSizeRequired, FLASH_SECTOR_SIZE) / FLASH_SECTOR_SIZE;
-
-    sector_addr = xip_virt_to_phys(MODEL_XIP_ADDR);           // e.g. 0x00200000
-
-    xprintf("Erasing %d sectors from 0x%08x to cover %lu bytes\n",
-    		sectors_needed, sector_addr, (unsigned long)flashSizeRequired);
-
-    // Disable XIP before erase
-    xip_enable_XIP(false);
-
-    if (xSemaphoreTake(xSPIMutex, portMAX_DELAY) != pdTRUE)  {
-        xprintf("Failed to take SPI mutex in xip_erase_model_flash_area()\n");
-        return -1;
-    }
-
-    for (uint16_t i = 0; i < sectors_needed; i++) {
-    	ret = hx_lib_spi_eeprom_erase_sector(spi_inst, sector_addr, FLASH_SECTOR);
-    	xprintf("  Erase sector %d addr 0x%08x -> result %d\n", i, sector_addr, ret);
-        if (ret != 0) {
-            xprintf("Failed to erase sector at address 0x%08x\n", sector_addr);
-            ret = -1;
-            break;
-        }
-        sector_addr += FLASH_SECTOR_SIZE;
-
-        // Force as task switch which will ensure the inactivity timeout doesn't happen.
-        // Will cause a call to vApplicationTaskSwitchedIn()
-        vTaskDelay(1);
-    }
-
-    xSemaphoreGive(xSPIMutex);
-    return ret;
-}
-
-/**
- * Write the model to the flash from the SD card file
- *
- * Step 1: open the model file and determine its size
- * Step 2: Erase flash to make space for the model and the meta data
- * Step 3: Write the filename info into the meta data area (and verify it)
- * Step 4: Write the model itself to the flash
- * Step 5: close the file and report the results and return status
- *
- * @param filename - such as '/MANIFEST/12V34.TFL'
- * @return 0 for success, -1 for fail
- */
-int xip_load_model_from_sd_to_flash(char *filename) {
-    FRESULT res;
-    FIL file;
-    UINT bytesRead;
-    DWORD fileSize;
-
-    uint32_t model_xip_addr;
-    uint32_t header_xip_addr;
-    uint32_t model_phys_addr;
-    uint32_t header_phys_addr;
-
-    uint32_t flash_address;
-    uint32_t totalBytesRead;
-    uint32_t write_size;
-
-
-    union {
-        uint8_t bytes[MODEL_XIP_INFO_SIZE];
-        uint32_t words[MODEL_XIP_INFO_SIZE / 4];
-    } fname_header, fname_verify;
-
-    // aligned buffers
-    uint8_t write_buf[256] __attribute__((aligned(4)));
-    uint8_t verify_buf[256] __attribute__((aligned(4)));
-
-    // Step 1: open the model file and determine its size
-
-    // If 'filename' includes a path, use it to open the file but only store the basename in flash header
-    const char *src_path = filename; // may include /Manifest/
-    const char *base = filename;     // basename part only
-    for (const char *p = filename; *p; ++p)  {
-        if (*p == '/' || *p == '\\')
-            base = p + 1; // advance past last separator
-    }
-
-    xprintf("load_model_from_sd_to_flash: src_path='%s' base='%s'\n", src_path, base);
-
-    if (xip_init_flash() != 0)  {
-        return -1;
-    }
-
-    res = f_open(&file, src_path, FA_READ);
-    if (res != FR_OK)  {
-        xprintf("Failed to open model file %s (error: %d)\n", src_path, res);
-        return -1;
-    }
-
-    fileSize = f_size(&file);
-    if (fileSize == 0)  {
-        xprintf("Model file is empty\n");
-        f_close(&file);
-        return -1;
-    }
-    xprintf("Model file size: %lu bytes\n", fileSize);
-
-    // Step 2: Erase flash to make space for the model and the meta data
-
-    if (xip_erase_model_flash_area(fileSize) != 0)  {
-        xprintf("Failed to erase flash for model\n");
-        f_close(&file);
-        return -1;
-    }
-
-    // Step 3: Write the filename info into the meta data area (and verify it)
-
-    model_xip_addr = MODEL_XIP_ADDR;
-    header_xip_addr = model_xip_addr - MODEL_XIP_INFO_SIZE;
-    model_phys_addr = xip_virt_to_phys(model_xip_addr);
-    header_phys_addr = xip_virt_to_phys(header_xip_addr);
-
-    // Take semaphore for header write
-    if (xSemaphoreTake(xSPIMutex, portMAX_DELAY) != pdTRUE)  {
-        xprintf("Failed to take SPI mutex for header write\n");
-        f_close(&file);
-        return -1;
-    }
-
-    // Ensure XIP is disabled
-    xip_enable_XIP(false);
-
-    memset(fname_header.bytes, 0, sizeof(fname_header.bytes));
-
-    strncpy((char *)fname_header.bytes, base, sizeof(fname_header.bytes) - 1);
-    xprintf("Writing filename header: %s\n", fname_header.bytes);
-
-    // write header (filename) via SPI (4 words)
-    int wr = hx_lib_spi_eeprom_word_write(spi_inst, header_phys_addr, fname_header.words, MODEL_XIP_INFO_SIZE);
-    if (wr != 0)  {
-        xprintf("Failed to write filename header to flash (err %d)\n", wr);
-        xSemaphoreGive(xSPIMutex);
-        f_close(&file);
-        return -1;
-    }
-
-    // small delay to let write finish
-    // TODO -why?
-    for (volatile int d = 0; d < 20000; ++d)  {
-        asm volatile("nop");
-    }
-
-    // verify header by reading back via SPI
-    memset(fname_verify.bytes, 0, sizeof(fname_verify.bytes));
-
-    if (hx_lib_spi_eeprom_word_read(spi_inst, header_phys_addr, fname_verify.words, MODEL_XIP_INFO_SIZE) != 0) {
-        xprintf("Failed to verify filename header in flash (read error)\n");
-        xSemaphoreGive(xSPIMutex);
-        f_close(&file);
-        return -1;
-    }
-
-    if (memcmp(fname_header.bytes, fname_verify.bytes, MODEL_XIP_INFO_SIZE) != 0)  {
-        xprintf("Filename header verify FAIL (SPI)\n");
-        xSemaphoreGive(xSPIMutex);
-        f_close(&file);
-        return -1;
-    }
-
-    xSemaphoreGive(xSPIMutex);
-
-    // Step 4: Write the model itself to the flash
-
-    // Write model data at model_phys_addr
-    flash_address = model_phys_addr;
-    totalBytesRead = 0;
-
-    while (totalBytesRead < fileSize) {
-        memset(write_buf, 0, sizeof(write_buf));
-        res = f_read(&file, write_buf, sizeof(write_buf), &bytesRead);
-        if (res != FR_OK || bytesRead == 0)  {
-            break;
-        }
-        write_size = (bytesRead + 3) & ~3U;
-        if (write_size > bytesRead)  {
-            memset(write_buf + bytesRead, 0, write_size - bytesRead);
-        }
-
-        // Take semaphore for each chunk write
-        if (xSemaphoreTake(xSPIMutex, portMAX_DELAY) != pdTRUE)  {
-            xprintf("Failed to take SPI mutex for chunk write\n");
-            f_close(&file);
-            return -1;
-        }
-
-        // Ensure XIP is disabled
-        xip_enable_XIP(false);
-
-        xprintf("Writing %u bytes (aligned %u) to 0x%08lX\n",
-                (unsigned)bytesRead, (unsigned)write_size, (unsigned long)flash_address);
-        if (hx_lib_spi_eeprom_word_write(spi_inst,
-                                         flash_address,
-                                         (uint32_t *)write_buf,
-                                         write_size) != 0) {
-            xprintf("Flash write failed at 0x%08lX\n", flash_address);
-            xSemaphoreGive(xSPIMutex);
-            f_close(&file);
-            return -1;
-        }
-
-//        // TODO - Why a delay? In any event, use FreeRTOS delay
-//        for (volatile int d = 0; d < 10000; ++d)  {
-//            asm volatile("nop");
-//        }
-
-        // Force as task switch which will ensure the inactivity timeout doesn't happen.
-        // Will cause a call to vApplicationTaskSwitchedIn()
-        vTaskDelay(1);
-
-        memset(verify_buf, 0, write_size);
-        if (hx_lib_spi_eeprom_word_read(spi_inst,
-                                        flash_address,
-                                        (uint32_t *)verify_buf,
-                                        write_size) != 0)  {
-            xprintf("Flash verify read failed at 0x%08lX\n", flash_address);
-            xSemaphoreGive(xSPIMutex);
-            f_close(&file);
-            return -1;
-        }
-        if (memcmp(write_buf, verify_buf, write_size) != 0) {
-            xprintf("Verify FAIL at 0x%08lX\n", flash_address);
-            xSemaphoreGive(xSPIMutex);
-            f_close(&file);
-            return -1;
-        }
-
-        xSemaphoreGive(xSPIMutex);
-
-        flash_address += write_size;
-        totalBytesRead += bytesRead;
-    }
-
-    // Step 5: close the file and report the results and return
-
-    f_close(&file);
-
-    if (totalBytesRead == fileSize) {
-        xprintf("Model successfully written to physical 0x%08lX (%lu bytes)\n",
-                (unsigned long)MODEL_FLASH_ADDR, (unsigned long)fileSize);
-#ifdef OLD
-        // Persist the model info to flash so it survives reboots
-        write_persisted_model_info();
-#endif // OLD
-
-        // Re-enable XIP mode after all writes complete
-        if (xip_enable_XIP(true)) {
-        	xprintf("XIP mode re-enabled\n");
-        	return 0;
-        }
-        else {
-        	return -1;
-        }
-    }
-
-    xprintf("Incomplete write: %lu/%lu bytes\n",
-            (unsigned long)totalBytesRead, (unsigned long)fileSize);
-    return -1;
-}
-
-/**
- * checks if a model with a specified name is in the XIP flash
- *
- * @param filename - name of model to look for
- * @return true if present
- */
-static bool xip_is_model_in_flash(char *filename) {
-    int ret = 0;
-    ModelMetaData metaDataRam;
-    uint32_t meta_physical_addr;
-    uint8_t maxLabels;
-
-    if (xip_init_flash() != 0)  {
-        return false;
-    }
-
-    // debug_flash_status();	// print debug info
-
-    // Round up meta size and model size to be a multiple of 4 bytes (word aligned)
-    // Actually, no need to round up.
-    uint32_t meta_length = xip_align_up(sizeof(ModelMetaData), 4);
-
-    //Sanity checks
-    if ((meta_length & 0x3) != 0) {
-        return -1;  // size not word-aligned (should never happen)
-    }
-
-    // The meta data is written at the start of the XIP model area
-    meta_physical_addr = xip_virt_to_phys(MODEL_XIP_ADDR);           // e.g. 0x00200000
-
-    // Ensure XIP is disabled before SPI access
-    xip_enable_XIP(false);
-
-    // Take semaphore
-    if (xSemaphoreTake(xSPIMutex, portMAX_DELAY) != pdTRUE) {
-        xprintf("Failed to take SPI mutex for is_model_in_flash\n");
-        return false;
-    }
-
-    // Read the flash to the local meta data structure
-    ret = hx_lib_spi_eeprom_word_read(spi_inst, meta_physical_addr, (uint32_t *)&metaDataRam, meta_length);
-
-    // TODO check all this giving and taking...
-    xSemaphoreGive(xSPIMutex);
-
-    if (ret != 0) {
-    	return false;
-    }
-
-	// Check the magic word
-	if (metaDataRam.magic != LABEL_MAGIC) {
-		xprintf("Missing signature 0x%08x\n", LABEL_MAGIC);
-		return false;
-	}
-
-	// Enable XIP mode to permit memory access to the XIP Flash
-    if (xip_enable_XIP(true)) {
-		xprintf("XIP mode re-enabled\n");
-    }
-    else {
-    	// error - should not happen
-    	return false;
-    }
-
-    // Only print this info at cold boot (for testing)
-    if (coldBoot) {
-    	XP_LT_GREY;
-    	xprintf("Meta data found in flash\n");
-    	printf_x_printBuffer((const uint8_t *)&metaDataRam, sizeof(ModelMetaData));
-
-    	// Small delay for buffer to print
-    	vTaskDelay(pdMS_TO_TICKS(10));
-
-    	XP_WHITE;
-
-    	xprintf("Magic: 0x%08x Model '%s' has %d labels of %d bytes:\n",
-    			(int) metaDataRam.magic, metaDataRam.modelName,
-				metaDataRam.class_count, metaDataRam.label_len);
-
-    	maxLabels = metaDataRam.class_count;
-    	if (maxLabels > MAX_CLASSES) {
-    		maxLabels = MAX_CLASSES;
-    	}
-
-    	for (uint8_t i=0; i < maxLabels; i++) {
-    		xprintf("%d = '%s'\n", i, metaDataRam.labels[i]);
-    	}
-    }
-
-    // Compare filename
-    // Use the filename length of 13 bytes - IMAGEFILENAMELEN
-    if (strncmp(filename, metaDataRam.modelName , IMAGEFILENAMELEN) == 0) {
-    	xprintf("Model file name in flash is '%s'\n", filename);
-    	return true;
-    }
-    else {
-    	char flashName[IMAGEFILENAMELEN];
-    	strncpy(flashName, metaDataRam.modelName, IMAGEFILENAMELEN - 1);
-    	flashName[IMAGEFILENAMELEN - 1] = '\0';	// Be safe
-    	xprintf("Model file name in flash is '%s', not '%s'\n", flashName, filename);
-    	return false;
-    }
-
-}
-
-/**
- * checks if a model with a specified name is in the SD card
- *
- * Search for the fike in the /MANIFEST folder
- *
- * @param filename - name of model to look for
- * @return true if present
- */
-static bool xip_is_file_in_sd(char * filename) {
-	char manifest_path[10 + IMAGEFILENAMELEN];			// let's restrict this to "/12345678/" then the filename, so 10 + IMAGEFILENAMELEN
-    FILINFO fno;
-    FRESULT res;
-
-    if (fatfs_mounted()) {
-    	snprintf(manifest_path, sizeof(manifest_path), "%s/%s", CONFIG_DIR, filename);
-    	xprintf("Looking for '%s'\n", manifest_path);
-
-        res = f_stat(manifest_path, &fno);
-
-        return (res == FR_OK);
-    }
-    else {
-    	return false;
-    }
-}
-
-/**
- * Search for 'TFL3' at the right place in flash.
- *
- * Tested using SPI read of the hearder
- *
- * @return true of a model appears present
- */
-static bool xip_valid_model_in_flash(void) {
-	uint32_t model_physical_addr;
-    // The 'TLF3' string should be at spi_hdr[4] for 4 bytes
-    uint8_t spi_hdr[8] __attribute__((aligned(4))) = {0};
-
-    // Ensure XIP is disabled before SPI access
-    if (!xip_enable_XIP(false)) {
-    	xprintf("Failed to disable XIP %d\n");
-    	xSemaphoreGive(xSPIMutex);
-    	return false;
-    }
-
-    // Take semaphore
-    if (xSemaphoreTake(xSPIMutex, portMAX_DELAY) != pdTRUE)  {
-        xprintf("Failed to take SPI mutex for valid_model_in_flash()\n");
-        return false;
-    }
-
-    // The meta data is written at the start of the XIP model area
-    // The model starts on a 16-byte boundary beyond the meta data
-    model_physical_addr = xip_virt_to_phys(MODEL_XIP_ADDR) + xip_align_up(sizeof(ModelMetaData), 16);
-
-    if (hx_lib_spi_eeprom_word_read(spi_inst, model_physical_addr, (uint32_t *)spi_hdr, sizeof(spi_hdr)) != 0) {
-        xprintf("Failed to read SPI header\n");
-        xSemaphoreGive(xSPIMutex);
-        return false;
-    }
-
-    if (memcmp(&spi_hdr[4], "TFL3", 4) != 0) {
-        // no match
-        xprintf("No valid TFLite model in flash\n");
-        xSemaphoreGive(xSPIMutex);
-        return false;
-    }
-
-    xSemaphoreGive(xSPIMutex);
-    return true;
-}
-
-// Modified load function - test BOTH SPI and XIP access
-/**
- * checks if a plausible model exists in the flash memory.
- *
- * If so, the XIP access is enabled and a pointer to the start of teh model is returned.
- *
- * @return pointer to the model, or null
+ * Thin C++ wrapper around xip_get_model_xip_address() — obtains the
+ * validated virtual address then calls tflite::GetModel().
  */
 static const tflite::Model *xip_load_model_from_flash(void) {
-	// Virtual address of the start of the TFTL model
-	uint32_t model_xip_address;
-
-	if (xip_init_flash() != 0) {
-		xprintf("Flash init failed\n");
-		return nullptr;
-	}
-
-	// Step 1: Read model via SPI to confirm data exists
-
-	// Check if we have valid TFLite header via SPI
-	if (!xip_valid_model_in_flash()) {
-		xprintf("No valid TFLite model in flash\n");
-		return nullptr;
-	}
-
-	// Step 2: Enable XIP mode for memory-mapped access
-
-	if (xip_enable_XIP(true)) {
-		// The model starts on a 16-byte boundary beyond the meta data
-		model_xip_address = MODEL_XIP_ADDR + xip_align_up(sizeof(ModelMetaData), 16);
-		return tflite::GetModel((const void *) model_xip_address);
-	}
-	else {
-		return nullptr;
-	}
-}
-
-/**
- * Enable or disable XIP
- *
- * XIP must be enabled before accessing the model via memory references
- *
- * @param enable - true to enable, false to disable
- * @return true if OK
- */
-static bool xip_enable_XIP(bool enable) {
-    // Take semaphore
-    if (xSemaphoreTake(xSPIMutex, portMAX_DELAY) != pdTRUE)  {
-        xprintf("Failed to take SPI mutex in xip_enable_XIP()\n");
-        return false;
-    }
-
-    if (hx_lib_spi_eeprom_enable_XIP(spi_inst, enable, FLASH_QUAD, true) != 0) {
-        xprintf("Failed to enable XIP mode\n");
-        xSemaphoreGive(xSPIMutex);
-        return false;
-    }
-
-    xSemaphoreGive(xSPIMutex);
-    return true;
-}
-
-/********************************** Functions to support saving and fetching meta data from flash ********/
-
-/**
- * Copies the meta object to flash
- *
- * The copy uses SPI to the physical address 0x00200000
- */
-int32_t xip_write_meta_data_to_flash(ModelMetaData * metaDataRam) {
-    uint32_t meta_physical_addr;
-    int32_t res;
-
-    uint32_t numBytes = sizeof(ModelMetaData);
-
-    /* Sanity checks */
-    if ((numBytes & 0x3) != 0) {
-        return -1;  // size not word-aligned (should never happen)
-    }
-
-    // The meta data is written at the start of the XIP model area
-    meta_physical_addr = xip_virt_to_phys(MODEL_XIP_ADDR);           // e.g. 0x00200000
-
-    if ((meta_physical_addr & 0x3) != 0) {
-        return -2;  // address not word-aligned
-    }
-
-    // Disable XIP before SPI write
-    xip_enable_XIP(false);
-    res = hx_lib_spi_eeprom_word_write(spi_inst, meta_physical_addr, (uint32_t *)metaDataRam, numBytes);
-
-    if (res !=0) {
-    	xprintf("Failed to write meta data %d\n", res);
-    }
-    return res;
-}
-
-/**
- * Copies model file from the SD card to the flash memory
- *
- * c.f. load_model_from_sd_to_flash()
- *
- * Step 1: allocate buffers
- * Step 2: open the model file and determine its size
- * Step 3: Erase flash to make space for the model and the meta data
- * Step 4: Write the model itself to the flash
- * Step 5: close the file and report the results and return status
- *
- * @param filename - such as '12V34.TFL'
- * @return true on success
- */
-static bool xip_copy_model_from_sd_to_flash(char * filename) {
-    FRESULT res;
-    FIL file;
-    UINT bytesRead;
-    DWORD fileSize;
-
-    uint32_t flash_address;
-    uint32_t totalBytesRead;
-    uint32_t write_size;
-    int32_t result;
-    uint32_t flashSizeRequired;
-
-	char manifest_path[10 + IMAGEFILENAMELEN];			// let's restrict this to "/12345678/" then the filename, so 10 + IMAGEFILENAMELEN
-
-	// Get absolute path to the SD card file
-	snprintf(manifest_path, sizeof(manifest_path), "%s/%s", CONFIG_DIR, filename);
-
-    if (xip_init_flash() != 0)  {
-    	return false;
-    }
-
-    // Step 1: allocate buffers
-
-    // aligned buffers taken from the heap: prevent stack buffer overflow
-    uint8_t *write_buf  = static_cast<uint8_t *>(pvPortMalloc(FILE_CHUNK_SIZE));
-    uint8_t *verify_buf = static_cast<uint8_t *>(pvPortMalloc(FILE_CHUNK_SIZE));
-
-    if (!write_buf || !verify_buf) {
-        vPortFree(write_buf);
-        vPortFree(verify_buf);
-        xprintf("Failed to allocate write_buf[], verifybuf[]\n");
-        return false;
-    }
-
-// Step 2: open the model file and determine its size
-
-    res = f_open(&file, manifest_path, FA_READ);
-    if (res != FR_OK)  {
-        xprintf("Failed to open model file %s (error: %d)\n", manifest_path, res);
-        vPortFree(write_buf);
-        vPortFree(verify_buf);
-        return false;
-    }
-
-    fileSize = f_size(&file);
-    if (fileSize == 0)  {
-        xprintf("Model file is empty\n");
-        f_close(&file);
-        vPortFree(write_buf);
-        vPortFree(verify_buf);
-        return false;
-    }
-
-    xprintf("Model file size: %lu bytes\n", fileSize);
-
-// Step 2: Erase flash to make space for the model and the meta data
-    // The model data must be pushed out to align on a 16 byte boundary
-    flashSizeRequired = xip_align_up(sizeof(ModelMetaData), 16) + fileSize;
-
-    if (xip_erase_model_flash_area(flashSizeRequired ) != 0)  {
-        xprintf("Failed to erase flash for model\n");
-        f_close(&file);
-        vPortFree(write_buf);
-        vPortFree(verify_buf);
-        return false;
-    }
-
-// Step 3: Write the model itself to the flash
-
-    // The meta data is written at the start of the XIP model area
-    // The model starts on a 16-byte boundary beyond the meta data
-    flash_address = xip_virt_to_phys(MODEL_XIP_ADDR) + xip_align_up(sizeof(ModelMetaData), 16);
-
-    xprintf("Writing model to 0x%08x\n", flash_address);
-
-    totalBytesRead = 0;
-
-    // Ensure XIP is disabled
-    xip_enable_XIP(false);
-
-    // Take semaphore for each chunk write????????????????
-    if (xSemaphoreTake(xSPIMutex, portMAX_DELAY) != pdTRUE)  {
-        xprintf("Failed to take SPI mutex for chunk write\n");
-        f_close(&file);
-        vPortFree(write_buf);
-        vPortFree(verify_buf);
-        return false;
-    }
-
-
-    while (1) {
-    	// Clear the buffer
-        memset(write_buf, 0, FILE_CHUNK_SIZE);
-
-    	res = f_read(&file, write_buf, FILE_CHUNK_SIZE, &bytesRead);
-    	if (res != FR_OK) {
-    		// handle error ;
-    	    f_close(&file);
-    	    xSemaphoreGive(xSPIMutex);
-    	    vPortFree(write_buf);
-    	    vPortFree(verify_buf);
-    	    return false;
-    	}
-    	if (bytesRead == 0) {
-    		// EOF
-    		break;
-    	}
-
-    	// Round up to a number of 32-bit words
-    	write_size = xip_align_up(bytesRead, 4);
-
-        xprintf("Writing %d bytes (aligned %d) to 0x%08x\n", bytesRead, write_size, (unsigned long)flash_address);
-        result = hx_lib_spi_eeprom_word_write(spi_inst,
-                flash_address,
-                (uint32_t *)write_buf,
-                write_size);
-
-        if (result != 0) {
-            xprintf("Flash write failed at 0x%08\n", flash_address);
-    		// handle error ;
-    	    f_close(&file);
-    	    xSemaphoreGive(xSPIMutex);
-    	    vPortFree(write_buf);
-    	    vPortFree(verify_buf);
-            return false;
-        }
-
-        // verify the write...
-        memset(verify_buf, 0, write_size);
-        result = hx_lib_spi_eeprom_word_read(spi_inst,
-                flash_address,
-                (uint32_t *)verify_buf,
-                write_size);
-
-        if (result != 0)  {
-            xprintf("Flash verify read failed at 0x%08\n", flash_address);
-    	    f_close(&file);
-    	    xSemaphoreGive(xSPIMutex);
-    	    vPortFree(write_buf);
-    	    vPortFree(verify_buf);
-    	    return false;
-        }
-
-        if (memcmp(write_buf, verify_buf, write_size) != 0) {
-        	xprintf("Verify FAIL at 0x%08\n", flash_address);
-        	f_close(&file);
-        	xSemaphoreGive(xSPIMutex);
-        	vPortFree(write_buf);
-        	vPortFree(verify_buf);
-        	return false;
-        }
-
-        flash_address += write_size;
-        totalBytesRead += bytesRead;
-    }
-
-    // Step 5: close the file and report the results and return status
-
-    xSemaphoreGive(xSPIMutex);
-
-    f_close(&file);
-    vPortFree(write_buf);
-    vPortFree(verify_buf);
-
-    if (totalBytesRead == fileSize) {
-    	xprintf("Model successfully written to 0x%08x (%lu bytes)\n",
-    			(unsigned long)MODEL_FLASH_ADDR, (unsigned long)fileSize);
-
-    	// Re-enable XIP mode after all writes complete
-        xip_enable_XIP(true);
-
-    	return true;
-    }
-
-    xprintf("Incomplete write: %lu/%lu bytes\n",
-    		(unsigned long)totalBytesRead, (unsigned long)fileSize);
-    return false;
-}
-
-/**
- * Copies model labels from the SD card to the flash memory
- * Also other metadata.
- *
- * @param modelName = same as filename e.g. '12345V67.TFL'
- * @return true on success
- */
-static bool xip_copy_metadata_to_flash(char * modelName) {
-    ModelMetaData metaDataRam;
-    uint8_t numLabels;
-
-    // Initialise
-    memset(&metaDataRam, 0, sizeof(ModelMetaData));
-
-    metaDataRam.magic       = LABEL_MAGIC;
-    //metaDataRam.class_count = g_class_count;	// This is not yet known!
-    metaDataRam.label_len   = MAX_LABEL_LEN;
-
-    // Copy model name e.g. 1234V567.TFL - 8.3 format
-    strncpy(metaDataRam.modelName, modelName, MAX_MODEL_NAME_LEN - 1);
-
-    // Copy the labels into metaDataRam
-    numLabels = load_labels_from_manifest(modelName, metaDataRam.labels);
-
-    metaDataRam.class_count = numLabels;
-
-    // Optional CRC over everything except crc field
-    //
-    //	crc16_ccitt_generate(gWrite_buf, I2CFMT_PAYLOAD_OFFSET + length, &checksum);
-    //
-    //    model_labels_ram.crc =
-    //        crc32((uint8_t *)&model_labels_ram,
-    //              offsetof(FlashModelLabels, crc));
-
-    // test that the structure is OK
-	XP_LT_GREY;
-	xprintf("Built this metadata:\n");
-    printf_x_printBuffer((uint8_t *)&metaDataRam, sizeof(ModelMetaData));
-	XP_WHITE;
-
-	// Write metaDataRam to flash
-	xip_write_meta_data_to_flash(&metaDataRam);
-
-    // test that the structure is OK
-	// Enable XIP mode to permit memory access to the XIP Flash
-
-    if (xip_enable_XIP(true)) {
-    	xprintf("XIP mode re-enabled\n");
-    }
-    else {
-    	// error should not happen
-    }
-
-	XP_LT_GREY;
-	xprintf("Meta data now in flash:\n");
-    printf_x_printBuffer((const uint8_t *)metaDataFlash, sizeof(ModelMetaData));
-	XP_WHITE;
-
-	// is there a chance of returning false:
-	return true;
+	uint32_t addr = xip_get_model_xip_address();
+	return addr ? tflite::GetModel((const void *)addr) : nullptr;
 }
 
 /********************************** Public Functions  *************************************/
@@ -1665,8 +619,7 @@ static bool xip_copy_metadata_to_flash(char * modelName) {
  *	@return 0 for OK, -1 if no NN is used
  */
 int cv_init(bool security_enable, bool privilege_enable, uint16_t project_id, uint16_t deploy_version, APP_WAKE_REASON_E woken) {
-	char filename[IMAGEFILENAMELEN];	// for 8.3 this is 13, including the training \0
-	uint32_t model_xip_address;
+	char filename[MAX_MODEL_NAME_LEN];	// for 8.3 this is 13, including the trailing \0
 
 	// Enforce clean state
 	cv_deinit();
@@ -1698,7 +651,7 @@ int cv_init(bool security_enable, bool privilege_enable, uint16_t project_id, ui
 	xprintf("Looking for model '%s' in flash or SD card\n", filename);
 
 	// Option 1: named model is in flash
-	if (xip_is_model_in_flash(filename)) {
+	if (xip_is_model_in_flash(filename, coldBoot)) {
 		xprintf("Flash already contains model '%s'; loading from flash.\n", filename);
 		modelUsed = xip_load_model_from_flash();
 	}
@@ -1706,19 +659,12 @@ int cv_init(bool security_enable, bool privilege_enable, uint16_t project_id, ui
 	else if (xip_is_file_in_sd(filename)) {
 		modelUsed = xip_load_model_from_sd(filename);
 	}
-	// Option 3: any model is in flash
-	else if(xip_valid_model_in_flash()) {
-		// Return any valid model
-	    // The model starts on a 16-byte boundary beyond the meta data
+	// Option 3: any model is in flash (not the named one, but something usable)
+	else if (xip_valid_model_in_flash()) {
 		xprintf("Found another valid model\n");
-		if (xip_enable_XIP(true)) {
-			// The model starts on a 16-byte boundary beyond the meta data
-			model_xip_address = MODEL_XIP_ADDR + xip_align_up(sizeof(ModelMetaData), 16);
-			modelUsed = tflite::GetModel((const void *) model_xip_address);
-		}
-		else {
-			// Only happens with error
-			xprintf("Error with xip_enable_XIP() \n");
+		modelUsed = xip_load_model_from_flash();
+		if (!modelUsed) {
+			xprintf("Error loading model from flash\n");
 			return -1;
 		}
 	}
@@ -1907,6 +853,7 @@ void cv_newModel(uint16_t project_id, uint16_t deploy_version) {
 		xprintf("Failed to send message 0x%04x\n", send_msg.msg_event);
 	}
 }
+
 // Robust deinit: safely release interpreter and tensor resources before model reloads
 int cv_deinit(void) {
 
@@ -1995,8 +942,6 @@ TfLiteStatus cv_run(int8_t *outCategories, uint8_t *categoriesCount) {
                 input->data.int8,
                 SC(app_get_raw_width(), input_width),
                 SC(app_get_raw_height(), input_height));
-
-    //xprintf("Image rescaled and loaded into input tensor.\n");
 
     TfLiteStatus invoke_status = interpreter->Invoke();
     xprintf("Model invoked.\n");
@@ -2091,10 +1036,6 @@ static void outputAsPercentage(TfLiteTensor *output) {
             confidence_frac = 0; // Handle rounding edge cases
         }
 
-        //const char *label = (g_labels_loaded && i < g_label_count) ? g_labels[i] : nullptr;
-//        const char *label = (g_labels_loaded && (i < metaDataFlash->class_count))
-//        		? metaDataFlash->labels[i] : nullptr;
-
         const char * label = metaDataFlash->labels[i];
 
         // Store confidence data for EXIF (if within bounds)
@@ -2179,4 +1120,3 @@ const char * cv_getLabel(uint8_t index) {
 }
 
 #endif // USE_PERCENTAGE
-
